+++ tests/test_procrunner.py
@@ -47,7 +47,7 @@ def test_run_command_aborts_after_timeout(
     task = ["___"]
 
     with pytest.raises(RuntimeError):
-        procrunner.run(task, timeout=-1, raise_timeout_exception=True)
+        procrunner.run(task, timeout=-1)
 
     assert mock_subprocess.Popen.called
     assert mock_process.terminate.called
@@ -81,7 +81,6 @@ def test_run_command_runs_command_and_directs_pipelines(
         callback_stdout=mock.sentinel.callback_stdout,
         callback_stderr=mock.sentinel.callback_stderr,
         working_directory=pathlib.Path("somecwd"),
-        raise_timeout_exception=True,
     )
 
     assert mock_subprocess.Popen.called
@@ -122,7 +121,7 @@ def test_run_command_runs_command_and_directs_pipelines(
 def test_default_process_environment_is_parent_environment(mock_subprocess):
     mock_subprocess.Popen.side_effect = NotImplementedError()  # cut calls short
     with pytest.raises(NotImplementedError):
-        procrunner.run([mock.Mock()], timeout=-1, raise_timeout_exception=True)
+        procrunner.run([mock.Mock()], timeout=-1)
     assert mock_subprocess.Popen.call_args[1]["env"] == os.environ
 
 
@@ -136,7 +135,6 @@ def test_pass_custom_environment_to_process(mock_subprocess):
             [mock.Mock()],
             timeout=-1,
             environment=copy.copy(mock_env),
-            raise_timeout_exception=True,
         )
     assert mock_subprocess.Popen.call_args[1]["env"] == mock_env
 
@@ -153,7 +151,6 @@ def test_pass_custom_environment_to_process_and_add_another_value(mock_subproces
             timeout=-1,
             environment=copy.copy(mock_env1),
             environment_override=copy.copy(mock_env2),
-            raise_timeout_exception=True,
         )
     mock_env_sum = copy.copy(mock_env1)
     mock_env_sum.update({key: str(mock_env2[key]) for key in mock_env2})
@@ -169,7 +166,6 @@ def test_use_default_process_environment_and_add_another_value(mock_subprocess):
             [mock.Mock()],
             timeout=-1,
             environment_override=copy.copy(mock_env2),
-            raise_timeout_exception=True,
         )
     random_environment_variable = list(os.environ)[0]
     if random_environment_variable == list(mock_env2)[0]:
@@ -199,7 +195,6 @@ def test_use_default_process_environment_and_override_a_value(mock_subprocess):
             environment_override={
                 random_environment_variable: "X" + random_environment_value
             },
-            raise_timeout_exception=True,
         )
     assert (
         mock_subprocess.Popen.call_args[1]["env"][random_environment_variable]